+++ packages/blaze-sdk/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @blaze-cardano/sdk
 
+## 0.1.20
+
+### Patch Changes
+
+- Updated dependencies [97ca100]
+  - @blaze-cardano/tx@0.5.3
+  - @blaze-cardano/uplc@0.1.14
+  - @blaze-cardano/wallet@0.1.39
+
+## 0.1.19
+
+### Patch Changes
+
+- Updated dependencies [42257b8]
+  - @blaze-cardano/tx@0.5.2
+  - @blaze-cardano/uplc@0.1.13
+  - @blaze-cardano/wallet@0.1.38
+
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies [ce33e70]
+  - @blaze-cardano/core@0.4.2
+  - @blaze-cardano/query@0.2.11
+  - @blaze-cardano/tx@0.5.1
+  - @blaze-cardano/uplc@0.1.12
+  - @blaze-cardano/wallet@0.1.37
+
 ## 0.1.17
 
 ### Patch Changes