+++ packages/blaze-wallet/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @blaze-cardano/wallet
 
+## 0.1.39
+
+### Patch Changes
+
+- Updated dependencies [97ca100]
+  - @blaze-cardano/tx@0.5.3
+
+## 0.1.38
+
+### Patch Changes
+
+- Updated dependencies [42257b8]
+  - @blaze-cardano/tx@0.5.2
+
+## 0.1.37
+
+### Patch Changes
+
+- Updated dependencies [ce33e70]
+  - @blaze-cardano/core@0.4.2
+  - @blaze-cardano/query@0.2.11
+  - @blaze-cardano/tx@0.5.1
+
 ## 0.1.36
 
 ### Patch Changes